Refactor & functionality to stop inactive sessions
